mutualfunds-web/backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
python-multipart
yfinance
openai
requests
numpy
pandas
rapidfuzz
scikit-learn
PyMuPDF
